Transpiling! Transpiling?
It took some work to get AS3JS to sucessfully transpile the app (I had to modify some of the source for AS3JS, haha), but I got it to work! I think the transpiled source is just as hard to read though, so I'm not sure how much it will really help me out in the rewrite process. I think I'm gonna have to just rewrite it all myself, and just sift through the original packages as much as I can. I had to remove some of the packages to prevent stack overflow errors in Node while parsing the source, so the compiled JS file is missing the source for the MinecraftGuy, BaseScrollPane, ComboBox, ScrollBar, and Slider packages. I think they should still be referenced in there, but I had to remove their actual sources to keep the transpiler from erroring out. This is a real heck of a project, when the transpiler bugs out on ya, hahaha! This is gonna be a wild, but fun ride :D

For AS3JS, I had to fix a directory loading issue, where an asterisk was at the end of the path for some reason. I did this by adding this call right after the other replace statements on line 242.

```js
/* Windows */
tmp = tmp.replace(/\\\*/g,'\\');

/* Unix */
tmp = tmp.replace(/\/\*/g,'/');
```