Node Source Sorter!
Started working on a Node script that will help me organize the original source of the project! It took me a good while to get this working, and I probably could have had a lot of it organized by now, but that's not what programmers are supposed to do! The goal is to write programs to do the work for you, rather than do the work.
This initial function is based a lot off of the `getFileSystemHandlesFromDirectory` function in my Librar.io project. This time it's built to use Node's fs module instead of the File System Access API in the browser. (Note: I can't believe how fast it performs in Node! I don't remember it being this quick in the browser. That's definitely piqued my interest in using Node for some kind of project in the future...)
I will be using this sorter function to eventually organize the files inside of the classic source folder so I can better navigate it to find things to use them in my own port of the project. There's too many folders and files at the moment for me to try and go through it with ease. I want to use my scripting knowledge to help ease the process of finding things myself. It's been super fun working with Node again! Haven't done too much with it yet except for things like Electron, NBT parsing, and Tauri. This time, I'm using Node as a tool to help with my own dev process! Super cool :D

Here's some of the articles I found that helped me through the dev process today. Some of them aren't related to the Node portion, but they gave me ideas for other things I could do later too.

With these, I thought it could be interesting if I made a ponyfill type of thing like the File System Access API feature set, but for Node! Kind of wacky, since you can already use the fs module, but sounds kind of cool to me. Here's those articles related to that:
https://developer.mozilla.org/en-US/docs/Web/API/FileSystemHandle
https://stackoverflow.com/questions/22156326/private-properties-in-javascript-es6-classes
https://nodejs.dev/learn/working-with-file-descriptors-in-nodejs
(These links were to get an idea on how to implement the `getFile()` function on a ponyfill `FileSystemFileHandle` class for Node.)
https://stackoverflow.com/questions/46867517/how-to-read-file-with-async-await-properly
https://stackoverflow.com/questions/43231241/how-to-create-a-file-object-with-a-path-in-nodejs

The rest of the articles are for the file looping logic itself, which is what is present in `index.js`.
https://stackoverflow.com/questions/15630770/node-js-check-if-path-is-file-or-directory/70328065#70328065
https://coderrocketfuel.com/article/7-methods-for-working-with-directories-in-node-js
https://nodejs.org/api/fs.html#class-fsdirent
https://nodejs.org/api/path.html#pathresolvepaths
https://nodejs.org/api/path.html#pathjoinpaths
https://stackoverflow.com/questions/62837771/how-to-get-the-file-using-relative-path-argument-in-nodejs
https://stackoverflow.com/questions/53799385/how-can-i-convert-a-windows-path-to-posix-path-using-node-path

Sorry for putting tons of links here! Thought it would be good to document everything that I found along the way. Pretty much all of the Node things there were new to me, so these articles really helped me piece together things a lot!