Organizing the Classic Source: Pt. 2
Wrote another Node script to un-nest a lot of the extra folders that contained some of the app state renders.

I think these were mostly generated by the Flash decompiler tool, so I don't think all of these were necessarily all part of the original SVF. Either way, they are super helpful to get some of the missing resources that weren't decompiled directly (Like the Skincraft logo itself, that seems to only be in this folder).

Using Node for organizing file and folder structures is just blowing my mind! This is so powerful. You can really do anything you'd like with it. There are so many tools available to you in the fs module, it's insanely helpful! This would've taken ages to do myself, haha.